Get status widget a bit more organized
