+++ mac-os/ubersicht-widgets/ubersicht-widgets/status/status.css
